% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{runGadget}
\alias{runGadget}
\title{Run gadget}
\usage{
runGadget(app, port = getOption("shiny.port"),
  launch.browser = getOption("viewer", TRUE))
}
\arguments{
\item{app}{A \code{\link[shiny]{shinyApp}} object.}

\item{port}{The TCP port that the application should listen on. If the port
is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}
}
\description{
Runs a Shiny Gadget app. If running in RStudio, the gadget will be displayed
in the Viewer pane; if not, and the \code{viewer} option has not been set,
then the gadget will be opened in your system's browser.
}

